--- a/src/core/lv_obj.c
+++ b/src/core/lv_obj.c
@@ -35,10 +35,6 @@
     #include "../gpu/lv_gpu_nxp_pxp_osa.h"
 #endif
 
-#if LV_USE_GPU_SDL
-    #include "../gpu/lv_gpu_sdl.h"
-#endif
-
 /*********************
  *      DEFINES
  *********************/
@@ -127,9 +123,6 @@ void lv_init(void)
     //        for(; ;) ;
     //    }
     //#endif
-    //#if LV_USE_GPU_SDL
-    //    lv_gpu_sdl_init();
-    //#endif
 
     _lv_obj_style_init();
     _lv_ll_init(&LV_GC_ROOT(_lv_disp_ll), sizeof(lv_disp_t));
@@ -186,13 +179,10 @@ void lv_init(void)
     LV_LOG_TRACE("finished");
 }
 
-#if LV_ENABLE_GC || !LV_MEM_CUSTOM || LV_USE_GPU_SDL
+#if LV_ENABLE_GC || !LV_MEM_CUSTOM
 
 void lv_deinit(void)
 {
-#if LV_USE_GPU_SDL
-    lv_gpu_sdl_deinit();
-#endif
     _lv_gc_clear_roots();
 
     lv_disp_set_default(NULL);