--- a/src/draw/sdl/lv_draw_sdl.c
+++ b/src/draw/sdl/lv_draw_sdl.c
@@ -64,13 +64,13 @@ void lv_draw_sdl_init_ctx(lv_disp_drv_t * disp_drv, lv_draw_ctx_t * draw_ctx)
     draw_ctx_sdl->base_draw.blend = lv_draw_sdl_blend;
     draw_ctx_sdl->internals = lv_mem_alloc(sizeof(lv_draw_sdl_context_internals_t));
     lv_memset_00(draw_ctx_sdl->internals, sizeof(lv_draw_sdl_context_internals_t));
-    lv_draw_sdl_texture_cache_init(draw_ctx_sdl->internals);
+    lv_draw_sdl_texture_cache_init(draw_ctx_sdl);
 }
 
 void lv_draw_sdl_deinit_ctx(lv_disp_drv_t * disp_drv, lv_draw_ctx_t * draw_ctx)
 {
     lv_draw_sdl_ctx_t *draw_ctx_sdl = (lv_draw_sdl_ctx_t *) draw_ctx;
-    lv_draw_sdl_texture_cache_deinit(draw_ctx_sdl->internals);
+    lv_draw_sdl_texture_cache_deinit(draw_ctx_sdl);
     lv_mem_free(draw_ctx_sdl->internals);
     _lv_draw_sdl_utils_deinit();
 }