--- a/src/draw/sdl/lv_draw_sdl_blend.c
+++ b/src/draw/sdl/lv_draw_sdl_blend.c
@@ -9,13 +9,13 @@
 
 #include "../../lv_conf_internal.h"
 
-#if LV_USE_DRAW_SDL
+#if LV_USE_GPU_SDL
 
 #include "lv_draw_sdl_texture_cache.h"
 #include "lv_draw_sdl_utils.h"
 #include "lv_draw_sdl_mask.h"
 
-#include LV_DRAW_SDL_INCLUDE_PATH
+#include LV_GPU_SDL_INCLUDE_PATH
 
 /*********************
  *      DEFINES
@@ -150,4 +150,4 @@ void blend_map(lv_draw_sdl_ctx_t * draw_ctx, const lv_draw_sw_blend_dsc_t * dsc)
     SDL_FreeSurface(surface);
 }
 
-#endif /*LV_USE_DRAW_SDL*/
+#endif /*LV_USE_GPU_SDL*/