### Description of the feature or fix

This update significantly improved drawing performance for SDL backend. Although it can't be reflected with benchmark, as it's already over 500fps, but it has over 50% higher framerate on webOS.

Before: 30fps
![Before: 30fps](https://user-images.githubusercontent.com/830358/145689995-0e889c75-6cd7-478c-bb91-701488b1e5c8.png)

After: > 50fps
![After: > 50fps](https://user-images.githubusercontent.com/830358/145689979-d65bd7c7-1517-4aab-9e70-83f8a401d15b.png)

Also, this PR improves rendering fidelity. For example, when drawing border with SIDE_RIGHT and SIDE_BOTTOM, previous implementation didn't draw anything.

Before:
![image](https://user-images.githubusercontent.com/830358/145690074-ed924b96-1ab8-46fd-9757-1f14af08c83f.png)

After:
![image](https://user-images.githubusercontent.com/830358/145690099-d7290fa5-e1f0-4e0e-a920-dfd34e3fe049.png)


### Checkpoints
- [ ] Follow the [styling guide](https://github.com/lvgl/lvgl/blob/master/docs/CODING_STYLE.md)
- [x] Run `code-format.py` from the `scripts` folder. [astyle](http://astyle.sourceforge.net/install.html) needs to be installed.
- [ ] Update the documentation