@mariotaku 2021-12-11
For failed builds, we'll need to change in https://github.com/lvgl/lv_binding_micropython/blob/e7ae68a0714fecee03f6c7e6842a5d1981cdba78/lv_conf.h#L168 .