@mariotaku 2021-12-14
@kisvegabor Changes contained in this PR now runs well. It may need some other changes, but before that I'd like to get this PR merged, and improve gradually.