@kisvegabor 2021-12-14
Wow, I'm always surprised by how fast you are :slightly_smiling_face: 

I mere it as you have asked! Thank you very much!